standard naming for components. move Treatment comp to own file
